Cleared out the warnings that came with the IBKR API

+++ src/ib_api_v9.76/ContractCondition.cpp
@@ -9,7 +9,7 @@
 std::string ContractCondition::toString() {
     std::string strContract = std::to_string(conId()) + "";
 
-    return std::to_string(type()) + " of " + strContract + OperatorCondition::toString();
+    return std::to_string((int)type()) + " of " + strContract + OperatorCondition::toString();
 }
 
 const char* ContractCondition::readExternal(const char* ptr, const char* endPtr) {
